Added axios interceptors to handle jwt tokens
1. Added axios interceptors to handle jwt tokens:
For requests, the interceptors will add the access token stored in the authslice. For response, when receiving Not Authorized response, the interceptors will try to get a new access token with the refreshToken stored in httpOnly cookie. Then retry the request with the new token. Any error will results in logout.
2. Replaced helper function fetch api with the axios calls, so that the interceptors can work.
3. From now on, the local state of authentication is managed by the redux store, while the refreshToken is managed by the backend. So logout requires both dispatch logout action and api calls to the backend.