Fixed interceptor issue to prevent race conditions
React renders twice, so it will send two request when a api is called in an event handler or useEffect. Hence, we will receive two 401 errors if our access token expires, but we only want to refresh the token once. That's why we use isRefreshing and refreshPromise to ensure that there is only one refresh request, and all other requests will wait for the same refresh promise and use the same token later.
If we don't do this, we will end up with multiple refresh requests being sent to the server. It will cause the first refresh token to be deleted from the database, resulting in the second request failing with a 401,
which fails to refresh the token as a result.